refactor: move collection error -> types
