+++ lib/collection/src/collection_builder/collection_loader.rs
@@ -1,17 +1,20 @@
-use crate::collection::{Collection, CollectionError};
+use std::fs::read_dir;
 use std::path::Path;
+use std::sync::Arc;
+
+use indicatif::ProgressBar;
 use tokio::runtime::Runtime;
+
+use segment::segment_constructor::segment_constructor::load_segment;
+
+use crate::collection::Collection;
+use crate::collection_builder::collection_builder::construct_collection;
+use crate::collection_builder::optimizers_builder::build_optimizers;
+use crate::config::CollectionConfig;
+use crate::operations::CollectionUpdateOperations;
+use crate::operations::types::CollectionError;
 use crate::segment_manager::holders::segment_holder::SegmentHolder;
 use crate::wal::SerdeWal;
-use crate::operations::CollectionUpdateOperations;
-use std::fs::{read_dir};
-use segment::segment_constructor::segment_constructor::load_segment;
-use crate::collection_builder::collection_builder::{construct_collection};
-use indicatif::ProgressBar;
-use crate::collection_builder::optimizers_builder::build_optimizers;
-use std::sync::Arc;
-use crate::config::CollectionConfig;
-
 
 pub fn load_collection(
     collection_path: &Path,
